Factor out shared test code for signatures
